% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_functions.R
\name{my_CI_upper}
\alias{my_CI_upper}
\title{My upper confidence interval function}
\usage{
my_CI_upper(value, se, prob = 0.95)
}
\arguments{
\item{value}{Input parameter for the function}

\item{se}{The standard error for the sample}

\item{prob}{The probability for the confidence interval, e.g. 0.95 for 95\%, based on the desired coverage.}
}
\value{
A numeric value for confidence interval
}
\description{
This function takes a mean and standard error and returns the value of upper limit of a normally distributed confidence interval.
}
\note{
This function will automatically convert to two-tailed probability, e.g. 0.95 to 0.975.
}
\examples{
#' # With a value of 100 and a known standard error of of 10, a 95\% interval is:
my_CI_lower(100, 10)

# A 99.8\% interval for the same is:
my_CI_lower(100, 10, 0.998)

}
