% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myinv.R
\name{myinv}
\alias{myinv}
\title{myinv}
\usage{
myinv(X, eps = 1e-12)
}
\arguments{
\item{X}{is the covriance matrix}

\item{eps}{is the precision when calculating the precision matrix.}
}
\value{
the precision matrix
}
\description{
This function computes the precision matrix, when the covariance matrix is singular.
}
\examples{
set.seed(111)
blocksize = 5; rho = 0.9
cov_block = matrix(0, blocksize ,blocksize)
for ( i in 1:blocksize){
for (j in 1:blocksize){
if(i==j){
cov_block[i,j] =  1
}else if( abs(i-j)>0 ){
cov_block[i,j] =  rho^ (abs(i-j))
}
}
}
Y = MASS::mvrnorm( n = 50,  mu = rep(0, blocksize ), Sigma = cov_block )
COV = cov(Y)
myinv(COV)

}
