% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage.R
\name{percentage}
\alias{percentage}
\title{Percentage function}
\usage{
percentage(numerator, denominator)
}
\arguments{
\item{numerator}{The numerator of the percentage will be divided by the denominator.}

\item{denominator}{The bottom of the fraction}
}
\value{
Returns a number or vector of numbers for percentages
}
\description{
Percentage function
}
\examples{
percentage(50, 100)


}
